Make the architecture a bit more granular
The game logic should be hidden in GameController now.

MessageHandler and receiver classes are for handling incoming
and outgoing messages.

RoomProvider is a class to handle all interactions with the
game room. This goes along with its callback classes for handling
events considering the room updates.

Finally, the CardProvider is for interfacing with the database
and accessing the cards.